Consolidate scramble logic so we can re-use it
Rename `index.tsx` files to the actual component names for ctrl+f ease
