+++ src/typescript/frontend/src/components/pages/home/components/main-card/MainCard.tsx
